feat: add tcea y valorentregado con mora

+++ ByteBank.API/Wallet/Domain/Models/Responses/WalletResponse.cs
@@ -11,9 +11,9 @@ public record WalletResponse
 
     public int Id { get; init; }
     public string NombreCartera { get; init; }
-    public decimal Tcea { get; init; }
-    public decimal ValorRecibido { get; init; }
-    public decimal ValorEntregado { get; init; }
+    public double Tcea { get; init; }
+    public double ValorRecibido { get; init; }
+    public double ValorEntregado { get; init; }
     public DateTime FechaDescuento { get; init; }
     public int PlazoOperacion { get; init; }
     public bool HasAssociatedBill { get; init; }
